+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ## Unreleased
 
+### Changes
+
+- Rename `navigation.processing` span to more expressive `Navigation dispatch to screen A mounted/navigation cancelled` ([#4423](https://github.com/getsentry/sentry-react-native/pull/4423))
+
 ### Dependencies
 
 - Bump CLI from v2.39.1 to v2.40.0 ([#4412](https://github.com/getsentry/sentry-react-native/pull/4412))